+++ Buildenator.Abstraction/BuildenatorConfigurationAttribute.cs
@@ -16,6 +16,7 @@ public sealed class BuildenatorConfigurationAttribute : Attribute
     /// <param name="implicitCast">Should the builder have implicit cast to the target type.</param>
     /// <param name="generateStaticPropertyForBuilderCreation">If you want to generate static property that will return a new builder instance.</param>
     /// <param name="initializeCollectionsWithEmpty">If true, collection fields will be initialized with empty collections in the constructor instead of null.</param>
+    /// <param name="useChildBuilders">If true, generates additional With methods that accept Func&lt;ChildBuilder, ChildBuilder&gt; for properties that have their own builders.</param>
     public BuildenatorConfigurationAttribute(
         string buildingMethodsPrefix = "With",
         bool generateDefaultBuildMethod = true,
@@ -23,7 +24,8 @@ public sealed class BuildenatorConfigurationAttribute : Attribute
         bool generateMethodsForUnreachableProperties = false,
         bool implicitCast = false,
         bool generateStaticPropertyForBuilderCreation = false,
-        bool initializeCollectionsWithEmpty = false)
+        bool initializeCollectionsWithEmpty = false,
+        bool useChildBuilders = false)
     {
         }
 }